Naming updates to domain.json
The target of a relationship should refer to a defined resource by its
name.  The name by default is the key of the resource unless you specify
the name property explicitly.  It is probably the cleanest to use the
name property (so good!) but then it also has to be used in
that (singular) form in the target.

This only tackles the issue for the specific view I have been targeting
and there is a good chance more of these cases exist in the domain.json.